This has everything I worked on for project Design
The Project Design.docx has organized everything I think I was supposed to submit.
The flowchart, class diagram and schematic are also provided bost as .png and as files in the software that created them. I have time tomorrow that if there is a major issue I can work on fixing them. Thank you for your time